% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\docType{data}
\name{CONUS_STATES}
\alias{CONUS_STATES}
\title{CONUS States}
\format{
An object of class \code{character} of length 49.

An object of class \code{character} of length 49.
}
\usage{
CONUS_STATES

CONUS_STATES
}
\description{
Vector of Continental United States state abbreviations

Vector of Continental United States state abbreviations
}
\keyword{datasets}
