% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\name{format_huc8}
\alias{format_huc8}
\title{Format HUC8 Codes}
\usage{
format_huc8(huc_codes)

format_huc8(huc_codes)
}
\arguments{
\item{huc_codes}{Character or numeric vector of HUC codes}
}
\value{
Character vector of properly formatted 8-digit HUC codes

Character vector of properly formatted 8-digit HUC codes
}
\description{
Add leading zeros to 7-digit HUC8 codes to make them 8-digit

Add leading zeros to 7-digit HUC8 codes to make them 8-digit
}
