% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\docType{data}
\name{KG_TO_TONS}
\alias{KG_TO_TONS}
\title{Conversion Factor: Kilograms to US Tons}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
KG_TO_TONS

KG_TO_TONS
}
\description{
Conversion Factor: Kilograms to US Tons

Conversion Factor: Kilograms to US Tons
}
\keyword{datasets}
