% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\docType{data}
\name{LBS_TO_KG}
\alias{LBS_TO_KG}
\title{Conversion Factor: Pounds to Kilograms}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
LBS_TO_KG

LBS_TO_KG
}
\description{
Conversion Factor: Pounds to Kilograms

Conversion Factor: Pounds to Kilograms
}
\keyword{datasets}
