% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\docType{data}
\name{LBS_TO_TONS}
\alias{LBS_TO_TONS}
\title{Conversion Factor: Pounds to US Tons}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
LBS_TO_TONS

LBS_TO_TONS
}
\description{
Conversion Factor: Pounds to US Tons

Conversion Factor: Pounds to US Tons
}
\keyword{datasets}
