% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\name{list_available_years}
\alias{list_available_years}
\title{List Available Built-in Years}
\usage{
list_available_years(scale = NULL)

list_available_years(scale = NULL)
}
\arguments{
\item{scale}{Character. Spatial scale (optional)}
}
\value{
Data frame with available years by data type

Data frame with available years by data type
}
\description{
Show available years for each data type

Show available years for each data type
}
