% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\docType{data}
\name{MANURESHED_CRS}
\alias{MANURESHED_CRS}
\title{Standard CRS for Manureshed Analysis}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
MANURESHED_CRS

MANURESHED_CRS
}
\description{
Albers Equal Area Conic projection (EPSG:5070)

Albers Equal Area Conic projection (EPSG:5070)
}
\keyword{datasets}
