% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_functions.R, R/utils.R
\docType{data}
\name{P2O5_TO_P}
\alias{P2O5_TO_P}
\title{Conversion Factor: P2O5 to P}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
P2O5_TO_P

P2O5_TO_P
}
\description{
Conversion Factor: P2O5 to P

Conversion Factor: P2O5 to P
}
\keyword{datasets}
